pyproject.toml
[tool.poetry]
name = "sun-angle"
version = "0.1.0"
description = ""
authors = ["roycehoe <royceinbox@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
pre-commit = "^3.2.2"
python-dotenv = "^1.0.0"
suncalc = "^0.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
